stream - custom collector

+++ src/main/java/com/siemens/model/Account.java
@@ -6,6 +6,7 @@ import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -14,6 +15,6 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = false)
 public class Account {
-    protected long runningTotals;
+    protected BigDecimal runningTotals;
     private LocalDate openDate;
 }
